requirements.txt
pyodbc~=4.0.32
python-telegram-bot
schedule